# openapi.yaml
openapi: 3.0.3
info:
  title: Pagaré Electrónico — API
  description: >-
    API del servicio de pagarés electrónicos. Todo lo que aquí figura lo expone **esta** plataforma; la
    API de BlockchainFUE que consumimos por debajo está documentada aparte, en openapi-bcf.yaml.


    Las validaciones siguen la Ley 19/1985, Cambiaria y del Cheque (LCCH), arts. 94-97 para el pagaré
    y las normas de la letra que le aplica el art. 96.


    **Alcance.** El registro se lleva en un libro mayor electrónico no cualificado (art. 3.52 del Reglamento
    (UE) 2024/1183): garantiza integridad y orden cronológico, pero no goza de la presunción de unicidad
    del art. 45 duodecies.2, reservada a los cualificados. El PDF que genera no es título ejecutivo: el
    juicio cambiario exige el documento original (STS, Sala 1ª, núm. 94/2014).
  version: 1.0.0
  license:
    name: Propietario — BAES Blockchain Labs (uso interno)
    url: https://github.com/BaesBlockchainLabs/pagare-electronico
servers:
- url: /
  description: La propia instancia del servicio
security:
- SesionCookie: []
...
paths:
...
  /api/pagares/firma:
    get:
      tags:
      - Pagarés
      summary: Estado de la firma cualificada del PDF
      description: >
        Informa de cómo va la firma del PDF de la última operación del pagaré. Consultarla es además lo
        que la recoge: si el firmante ya firmó, descarga el PDF, comprueba que el portal firmó
        exactamente el documento que se le mandó y completa la operación que estaba esperando —la
        entrega al beneficiario en una emisión, el cambio de titularidad en un endoso—.
      operationId: estadoFirmaPagare
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  activa:
                    type: boolean
                    description: Si esta instalación tiene la firma del PDF configurada. Cuando es
                      false, las operaciones se completan en el acto y no hay firma que esperar.
                  aviso:
                    type: string
                    description: Presente cuando la firma está bien pero la operación que esperaba
                      falló; se puede reintentar sin volver a firmar.
                  firma:
                    $ref: '#/components/schemas/FirmaPagare'
                  firmas:
                    type: array
                    description: Todas las firmas de la historia del pagaré, de la más antigua a la más
                      reciente. Cada una es una operación firmada.
                    items:
                      $ref: '#/components/schemas/FirmaPagare'
        4XX:
          $ref: '#/components/responses/Error'
    post:
      tags:
      - Pagarés
      summary: Volver a pedir la firma
      description: >
        Vuelve a pedir la firma de una operación cuya firma se quedó sin hacer. Sin esto un pagaré cuya
        firma falla queda inentregable para siempre: la emisión ya está en la cadena y no se puede
        deshacer, así que la única salida sería anularlo y volver a emitir con otro id.


        Se manda el mismo documento, no uno generado de nuevo, para que el firmante firme lo que se le
        pidió firmar. Sólo lo puede pedir quien tiene que firmar, o un administrador: reintentar manda
        un aviso al móvil de una persona. Si hay una firma en curso se niega, y si ya estaba firmada lo
        dice sin pedir nada.
      operationId: pedirFirmaDeNuevo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - id
              properties:
                id:
                  type: string
      responses:
        '200':
          description: Firma pedida de nuevo, o ya firmada
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SimpleOkResponse'
                - type: object
                  properties:
                    firma:
                      $ref: '#/components/schemas/FirmaPagare'
        4XX:
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'
...
  /api/pagares/firma/pdf:
    get:
      tags:
      - Pagarés
      summary: Descargar el PDF con la firma cualificada
      description: >
        El documento con valor probatorio autónomo: lleva una firma PAdES del prestador y un sello de
        tiempo RFC 3161, no sólo la firma ed25519 de la red. Sólo existe cuando la firma está
        completada.
      operationId: descargarPDFFirmado
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: El PDF firmado
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        4XX:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    SesionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sesión firmada, obtenida en /api/auth/login.
  responses:
    Error:
      description: >
        Error de la petición. El cuerpo incluye `{ ok: false, msg, errors? }`. Códigos típicos: 400 (validación/parámetros),
        404 (no encontrado), 502 (error de BlockchainFUE).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          default: false
        msg:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              campo:
                type: string
              mensaje:
                type: string
              articulo_lcch:
                type: string
                description: Artículo de la LCCH que no se cumple.
...
    FirmaPagare:
      type: object
      description: >
        La firma cualificada del PDF de una operación. Los dos hashes son lo que ata la firma al
        contenido: hash_original es el del PDF que se mandó a firmar, y el portal devuelve ese mismo
        documento byte a byte para poder cotejarlo.
      properties:
        estado:
          type: string
          enum:
          - ''
          - pendiente
          - firmada
          - fallida
        operacion:
          type: string
          enum:
          - emision
          - endoso
          - cesion
        referencia:
          type: string
          description: Identificador del envío en el portal de firma.
        hash_original:
          type: string
          description: SHA-256 del PDF que se mandó a firmar.
        hash_firmado:
          type: string
          description: SHA-256 del PDF firmado que devolvió el portal.
        motivo:
          type: string
          description: Por qué no se completó, en palabras del portal.
        msg:
          type: string
        creada_at:
          type: string
          format: date-time
        resuelta_at:
          type: string
          format: date-time
...
    SimpleOkResponse:
      type: object
      properties:
        ok:
          type: boolean
        msg:
          type: string
...